fix: save match team api update - 저장시 team 정보 업데이트 추가

+++ src/main/java/com/pickandlol/pickandlol/Model/ResponseTeamGetDTO.java
@@ -13,7 +13,7 @@ public class ResponseTeamGetDTO {
     Integer winCount;
     Integer loseCount;
     Integer totalGap;
-    Integer winRate;
+    Double winRate;
     Double kda;
     Integer killCount;
     Integer deathCount;
